ai-tutor-gemini/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
redis==5.0.4
